application/requirements.txt
numpy
opencv-python
Pillow
torch
